backend/app/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Environment & Config
python-dotenv==1.0.1

# Authentication & Security
python-jose==3.3.0
google-auth==2.28.0

# Vision & gRPC tooling
google-cloud-vision==3.4.4
protobuf<5,!=3.20.0,!=3.20.1,>=3.19.5
grpcio-tools==1.62.3

# Database ORM
sqlalchemy==2.0.27

# LLM & AI
groq==1.0.0

# Database & Vector Store
qdrant-client>=1.8.2

# File Processing & Document Handling
python-multipart==0.0.6
pymupdf==1.24.1
python-docx==0.8.11

tiktoken==0.8.0

# ML Frameworks
scikit-learn==1.4.2

# Pin NumPy to <2 to avoid ABI issues with compiled wheels (PyTorch, scikit-learn)
numpy<2

# Cloudinary for media management
cloudinary==1.41.0


#GOOGLE Vision
google-cloud-vision==3.4.4


# Transformers & ML
transformers>=4.44.2
torch>=2.2.0
Pillow==10.1.0
openai-whisper
sentence-transformers>=2.2.0